Extract non entry-point code to modules
