--- src/main/tv/codely/scala_http_api/user/domain/User.scala
@@ -1,7 +1,0 @@
-package tv.codely.scala_http_api.user.domain
-
-object User {
-  def apply(id: String, name: String): User = User(UserId(id), UserName(name))
-}
-
-case class User(id: UserId, name: UserName)
