--- src/main/tv/codely/scala_http_api/user/domain/UserId.scala
@@ -1,9 +1,0 @@
-package tv.codely.scala_http_api.user.domain
-
-import java.util.UUID
-
-object UserId {
-  def apply(value: String): UserId = UserId(UUID.fromString(value))
-}
-
-case class UserId(value: UUID)
