--- src/main/tv/codely/scala_http_api/user/domain/UserName.scala
@@ -1,3 +1,0 @@
-package tv.codely.scala_http_api.user.domain
-
-case class UserName(value: String)
