--- src/main/tv/codely/scala_http_api/video/domain/Video.scala
@@ -1,14 +1,0 @@
-package tv.codely.scala_http_api.video.domain
-
-import scala.concurrent.duration.Duration
-
-object Video {
-  def apply(id: String, title: String, duration: Duration, category: String): Video = Video(
-    VideoId(id),
-    VideoTitle(title),
-    VideoDuration(duration),
-    VideoCategory(category)
-  )
-}
-
-case class Video(id: VideoId, title: VideoTitle, duration: VideoDuration, category: VideoCategory)
