--- src/main/tv/codely/scala_http_api/video/domain/VideoDuration.scala
@@ -1,9 +1,0 @@
-package tv.codely.scala_http_api.video.domain
-
-import scala.concurrent.duration._
-
-object VideoDuration {
-  def apply(seconds: BigDecimal): VideoDuration = VideoDuration(seconds.longValue().seconds)
-}
-
-case class VideoDuration(value: Duration)
