--- src/main/tv/codely/scala_http_api/video/domain/VideoId.scala
@@ -1,9 +1,0 @@
-package tv.codely.scala_http_api.video.domain
-
-import java.util.UUID
-
-object VideoId {
-  def apply(value: String): VideoId = VideoId(UUID.fromString(value))
-}
-
-case class VideoId(value: UUID)
