--- src/main/tv/codely/scala_http_api/video/domain/VideoTitle.scala
@@ -1,3 +1,0 @@
-package tv.codely.scala_http_api.video.domain
-
-case class VideoTitle(value: String)
