+++ src/test/tv/codely/scala_http_api/module/stub/DurationStub.scala
@@ -1,4 +1,4 @@
-package tv.codely.scala_http_api.shared.infrastructure.stub
+package tv.codely.scala_http_api.module.stub
 
 import scala.concurrent.duration._
