+++ src/test/tv/codely/scala_http_api/module/stub/IntStub.scala
@@ -1,4 +1,4 @@
-package tv.codely.scala_http_api.shared.infrastructure.stub
+package tv.codely.scala_http_api.module.stub
 
 import scala.util.Random
