--- src/test/tv/codely/scala_http_api/user/infrastructure/stub/UserStub.scala
@@ -1,12 +1,0 @@
-package tv.codely.scala_http_api.user.infrastructure.stub
-
-import tv.codely.scala_http_api.user.domain.User
-
-object UserStub {
-  def apply(
-      id: String = UserIdStub.random.value.toString,
-      name: String = UserNameStub.random.value
-  ): User = User(id, name)
-
-  def random: User = apply()
-}
